Async emit
During load testing at whatnot, we discovered that emitting metrics
under load incurred latency. This was intentional, the genserver for
emitters was creating backparessure. However, we also enabled the max
dwell time feature, which would provide enough safety on its own.

This pr allows metrics to be emitted asynchronously when the library
is configured to use a max dwell time.